Add zsh and bash

+++ Makefile
@@ -14,7 +14,9 @@ UNLINK_TARGETS	+= unlink-$(1)
 .PHONY: link-$(1)
 link-$(1):
 	@for dotfile in $(2); do
-		ln -snfv "$$(DOTFILES_DIR)/$$$${dotfile}" "$$(HOME)/$$$${dotfile}" 
+		if [ -e "$$(DOTFILES_DIR)/$$$${dotfile}" ]; then
+			ln -snfv "$$(DOTFILES_DIR)/$$$${dotfile}" "$$(HOME)/$$$${dotfile}" 
+		fi
 	done
 unlink-$(1):
 	@for dotfile in $(2); do
